+++ src/main/java/com/allcleargateway/gateway/global/client/UserClient.java
@@ -1,14 +1,18 @@
-package com.allcleargateway.gateway;
+package com.allcleargateway.gateway.global.client;
 
-import com.allcleargateway.gateway.config.FeignConfig;
+import com.allcleargateway.gateway.domain.auth.dto.LoginDto;
+import com.allcleargateway.gateway.domain.auth.dto.TokenDto;
+import com.allcleargateway.gateway.global.config.FeignConfig;
+import com.allcleargateway.gateway.domain.member.dto.SignupDto;
+import com.allcleargateway.gateway.domain.member.dto.StudentDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
 
-@FeignClient(name = "waitingClient", url = "https://waiting.allclear-server.com:8081", configuration = FeignConfig.class)
-public interface WaitingClient {
+@FeignClient(name = "userClient", url = "https://user.allclear-server.com:8082", configuration = FeignConfig.class)
+public interface UserClient {
 
     @GetMapping(value = "/{relativeUrl}")
     ResponseEntity<String> forwardGetRequest(
@@ -43,4 +47,24 @@ public interface WaitingClient {
             @RequestBody String body,
             @PathVariable("relativeUrl") String relativeUrl
     );
+
+    //login을 위해 DTO와 DB에 있는 엔티티 비교
+    @PostMapping("/login")
+    ResponseEntity<StudentDto> forwardLoginPostRequest(
+            @RequestBody LoginDto.RequestDto body
+    );
+    
+    @PostMapping("/signup")
+    ResponseEntity<Boolean> forwardSignUpPostRequest(
+            @RequestBody SignupDto.RequestDto body
+    );
+
+    //로그인시 update된 refreshToken 저장
+    @PostMapping("/refreshToken")
+    ResponseEntity<Boolean> forwardRefreshTokenPostRequest(
+            @RequestBody TokenDto.UpdateRefreshTokenRequestDto body
+    );
+
+    @GetMapping("/student/info/{memberId}")
+    StudentDto forwardGetRequest(@PathVariable (name = "memberId")Long memberId);
 }
