+++ src/main/java/com/allcleargateway/gateway/global/client/WaitingClient.java
@@ -1,18 +1,14 @@
-package com.allcleargateway.gateway;
+package com.allcleargateway.gateway.global.client;
 
-import com.allcleargateway.gateway.auth.dto.LoginDto;
-import com.allcleargateway.gateway.auth.dto.TokenDto;
-import com.allcleargateway.gateway.config.FeignConfig;
-import com.allcleargateway.gateway.member.dto.SignupDto;
-import com.allcleargateway.gateway.member.dto.StudentDto;
+import com.allcleargateway.gateway.global.config.FeignConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
 
-@FeignClient(name = "userClient", url = "https://user.allclear-server.com:8082", configuration = FeignConfig.class)
-public interface UserClient {
+@FeignClient(name = "waitingClient", url = "https://waiting.allclear-server.com:8081", configuration = FeignConfig.class)
+public interface WaitingClient {
 
     @GetMapping(value = "/{relativeUrl}")
     ResponseEntity<String> forwardGetRequest(
@@ -47,24 +43,4 @@ public interface UserClient {
             @RequestBody String body,
             @PathVariable("relativeUrl") String relativeUrl
     );
-
-    //login을 위해 DTO와 DB에 있는 엔티티 비교
-    @PostMapping("/login")
-    ResponseEntity<StudentDto> forwardLoginPostRequest(
-            @RequestBody LoginDto.RequestDto body
-    );
-    
-    @PostMapping("/signup")
-    ResponseEntity<Boolean> forwardSignUpPostRequest(
-            @RequestBody SignupDto.RequestDto body
-    );
-
-    //로그인시 update된 refreshToken 저장
-    @PostMapping("/refreshToken")
-    ResponseEntity<Boolean> forwardRefreshTokenPostRequest(
-            @RequestBody TokenDto.UpdateRefreshTokenRequestDto body
-    );
-
-    @GetMapping("/student/info/{memberId}")
-    StudentDto forwardGetRequest(@PathVariable (name = "memberId")Long memberId);
 }
