+++ src/main/java/com/allcleargateway/gateway/global/config/SSLConfig.java
@@ -1,4 +1,4 @@
-package com.allcleargateway.gateway.config;
+package com.allcleargateway.gateway.global.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
